requirement.txt
boto3==1.37.6
python-dotenv==1.0.1
opencv-python==4.11.0
"numpy<2"
ultralytics
python-dotenv